refactor(expo): Extract GameCard component and improve UI layout
- Move GamesCard from games index to a separate GameCard component
- Adjust layout in games index page (SafeAreaView to View, height percentages)
- Add dynamic dropdown positioning in GamesDropDown
- Add direct link to game stats in dropdown menu
