+++ apps/expo/src/components/gamesDropDown.tsx
@@ -1,5 +1,8 @@
+import type { View } from "react-native";
 import * as React from "react";
+import { Dimensions } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { Link } from "expo-router";
 
 import type { RouterOutputs } from "~/utils/api";
 import { Button } from "~/components/ui/button";
@@ -26,6 +29,22 @@ export function GamesDropDown({
     left: 12,
     right: 12,
   };
+
+  const [dropdownSide, setDropdownSide] = React.useState<"top" | "bottom">(
+    "bottom",
+  );
+  const buttonRef = React.useRef<View>(null);
+
+  const checkPosition = React.useCallback(() => {
+    if (buttonRef.current) {
+      buttonRef.current.measure((_x, _y, _width, _height, _pageX, pageY) => {
+        const windowHeight = Dimensions.get("window").height;
+        const threshold = windowHeight * 0.7; // 70% of screen height
+        setDropdownSide(pageY > threshold ? "top" : "bottom");
+      });
+    }
+  }, []);
+
   const utils = api.useUtils();
   const deleteGame = api.game.deleteGame.useMutation({
     onSuccess: async () => {
@@ -47,7 +66,12 @@ export function GamesDropDown({
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Button variant="ghost" size="icon">
+        <Button
+          variant="ghost"
+          size="icon"
+          ref={buttonRef}
+          onLayout={checkPosition}
+        >
           <MoreVertical
             className="text-foreground"
             size={20}
@@ -55,18 +79,20 @@ export function GamesDropDown({
           />
         </Button>
       </DropdownMenuTrigger>
-      <DropdownMenuContent insets={contentInsets}>
+      <DropdownMenuContent insets={contentInsets} side={dropdownSide}>
         <DropdownMenuGroup>
           <DropdownMenuItem>
             <Text>Edit</Text>
           </DropdownMenuItem>
-          <DropdownMenuItem>
-            <Text>Stats</Text>
-          </DropdownMenuItem>
+          <Link href={`/games/${data.id}/stats`} asChild>
+            <DropdownMenuItem>
+              <Text>Stats</Text>
+            </DropdownMenuItem>
+          </Link>
           <DropdownMenuItem>
             <Text>Rules</Text>
           </DropdownMenuItem>
-          <DropdownMenuItem onPress={onDelete}>
+          <DropdownMenuItem onPress={() => onDelete()}>
             <Text className="text-destructive focus:bg-destructive/80 focus:text-destructive-foreground">
               Delete
             </Text>
